Add files via upload
This project is a comprehensive analysis of traffic accidents and vehicle involvement in the UK between 2010 and 2011, utilizing Python for data manipulation and statistical analysis. The team embarked on this analysis with the goals of understanding the relationships between various factors and accident severity, identifying patterns, and deriving insights that could inform safety improvements and preventive measures.